fix: mask_output fvd aggregation
If a output_parameters->stream_output->mask_output file was
configured, the flows, velocities, and depths written to disk are an
erroneous mixture of flows, velocities, and depths. For example, the
flow field for a given feature_id contains values for a mixture of
aggregated flows, velocities, and depths. Not the expected flows.

This patch corrects this behavior so the expected aggregated flows _and_
depths are written. However, I am not certain that velocities are
correctly aggregated. I observed NAN values returned for aggregated
velocities. The `optimized_v_data` subroutine needs to be checked to
ensure velocity values are correctly aggregated.